Fix comment, move invitations constants
